More progress
Finished getting the seed generation function in place as well as preparation for custom node js file creation for each user to talk to the blockchain.  I've never even heard of this kind of framework created so there will be bugs I'm sure.  Get ready for (shell_exec('$_SESSION['user_name'] .'_'. count($_REQUEST['address']) .'_iota.js') for node block spinnup and broadcast. Then, We done!!! (After the ability to save the blogposts to the MAM tangle and we can save the data timestamp.